refactor(engine): make active query dependencies acyclic
