+++ packages/column-live-view-engine/src/active-query-raw.test.ts
@@ -1,564 +1,22 @@
-import { fromStringUnsafe } from "effect/BigDecimal";
 import { describe, expect, it } from "@effect/vitest";
-import { Cause, Effect, Exit, Option, Result, Schema } from "effect";
+import { Effect, Option, Schema } from "effect";
 import {
-  acquireMaterializedQueryExecution,
   acquireRawQueryExecution,
-  activeStoreRawQueryExecutionCount,
-  createActiveQueryRegistry,
-  releaseMaterializedQueryExecution,
+  activeQueryTestInterface,
+  activeQueryTestMetadata,
   releaseRawQueryExecution,
-} from "./active-query";
-import { replaceRetainedMatchingEntryAtIndex } from "./active-raw-query";
-import { prepareRuntimeGroupedQuery } from "./grouped-query-compiler";
+} from "../test-harness/active-query-interface";
 import { prepareRuntimeRawQuery } from "./raw-query-compiler";
-import { makeQueryResultSemantics } from "./query-result-semantics";
 import {
   deleteTopicStoreRow,
   patchTopicStoreRow,
   publishTopicStoreRow,
   publishTopicStoreRows,
-  resetTopicStore,
   TopicStore,
 } from "./topic-store";
-import { topicStoreRawQueryMetadata, topicStoreReadModel } from "./topic-store-state";
 
 const invalidRow = (_topic: string, message: string): Error => new Error(message);
-const emptyResultSemantics = makeQueryResultSemantics([]);
-
-describe("column-live-view-engine active query execution", () => {
-  it("falls back when an indexed retained replacement points at a different key", () => {
-    const windowEntries = [
-      { key: "a", row: { id: "a", score: 2 } },
-      { key: "b", row: { id: "b", score: 1 } },
-    ];
-    const replaced = replaceRetainedMatchingEntryAtIndex(
-      windowEntries,
-      0,
-      "b",
-      { id: "b", score: 3 },
-      (left, right) => right.row.score - left.row.score,
-      50,
-    );
-
-    expect(replaced).toBeUndefined();
-    expect(windowEntries).toStrictEqual([
-      { key: "a", row: { id: "a", score: 2 } },
-      { key: "b", row: { id: "b", score: 1 } },
-    ]);
-  });
-
-  it.effect("reuses execution state for identical compiled raw queries", () =>
-    Effect.gen(function* () {
-      const rowSchema = Schema.Struct({
-        id: Schema.String,
-        status: Schema.String,
-        score: Schema.Number,
-        count: Schema.BigInt,
-        value: Schema.BigDecimal,
-      });
-      const store = new TopicStore("scores", rowSchema, "id", () => {});
-
-      yield* publishTopicStoreRow(
-        store,
-        {
-          id: "1",
-          status: "open",
-          score: 10,
-          count: 1n,
-          value: fromStringUnsafe("1.00"),
-        },
-        invalidRow,
-      );
-      yield* publishTopicStoreRow(
-        store,
-        {
-          id: "2",
-          status: "closed",
-          score: 20,
-          count: 2n,
-          value: fromStringUnsafe("2.00"),
-        },
-        invalidRow,
-      );
-
-      const compiled = yield* prepareRuntimeRawQuery("scores", topicStoreRawQueryMetadata(store), {
-        select: ["id", "score", "count", "value"],
-        where: {
-          status: "open",
-        },
-        orderBy: [
-          {
-            field: "score",
-            direction: "desc",
-          },
-        ],
-      });
-
-      const firstExecution = yield* acquireRawQueryExecution(topicStoreReadModel(store), compiled);
-      const secondExecution = yield* acquireRawQueryExecution(topicStoreReadModel(store), compiled);
-      expect(Object.isFrozen(firstExecution)).toBe(true);
-      expect(() => Object.assign(firstExecution, secondExecution)).toThrowError(TypeError);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(1);
-
-      const firstCursor = firstExecution.createCursor();
-      const secondCursor = secondExecution.createCursor();
-
-      const initialFirst = firstExecution.initial("query-a");
-      const initialSecond = secondExecution.initial("query-b");
-      expect(initialFirst.rows).toStrictEqual(initialSecond.rows);
-      expect(initialFirst.totalRows).toBe(1);
-      expect(initialFirst.keys).toStrictEqual(["1"]);
-
-      const beforePublishFirst = yield* firstExecution.next("query-a", firstCursor);
-      const beforePublishSecond = yield* secondExecution.next("query-b", secondCursor);
-      expect(beforePublishFirst._tag).toBe("None");
-      expect(beforePublishSecond._tag).toBe("None");
-
-      yield* publishTopicStoreRow(
-        store,
-        {
-          id: "3",
-          status: "open",
-          score: 5,
-          count: 3n,
-          value: fromStringUnsafe("3.00"),
-        },
-        invalidRow,
-      );
-
-      const afterPublishFirst = yield* firstExecution.next("query-a", firstCursor);
-      const afterPublishSecond = yield* secondExecution.next("query-b", secondCursor);
-      expect(afterPublishFirst._tag).toBe("Some");
-      expect(afterPublishSecond._tag).toBe("Some");
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), compiled);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(1);
-      const afterRefcountDecrement = yield* acquireRawQueryExecution(
-        topicStoreReadModel(store),
-        compiled,
-      );
-      expect(afterRefcountDecrement.initial("query-c").totalRows).toBe(2);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(1);
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), compiled);
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), compiled);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(0);
-
-      const afterRefcountExhausted = yield* acquireRawQueryExecution(
-        topicStoreReadModel(store),
-        compiled,
-      );
-      expect(afterRefcountExhausted.initial("query-d").totalRows).toBe(2);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(1);
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), compiled);
-    }),
-  );
-
-  it.effect("binds a storage projection once per raw execution lease", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "projection-bind-frequency",
-        Schema.Struct({
-          id: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      yield* publishTopicStoreRow(store, { id: "a", score: 1 }, invalidRow);
-      const readModel = topicStoreReadModel(store);
-      const storageProjection = Option.getOrThrow(
-        Option.fromUndefinedOr(readModel.storageProjection),
-      );
-      let projectionReads = 0;
-      const observedReadModel = {
-        ...readModel,
-        get storageProjection() {
-          projectionReads += 1;
-          return storageProjection;
-        },
-      };
-      const compiled = yield* prepareRuntimeRawQuery(
-        "projection-bind-frequency",
-        topicStoreRawQueryMetadata(store),
-        { select: ["id", "score"] },
-      );
-
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
-      const cursor = execution.createCursor();
-      expect(execution.initial("query").rows).toStrictEqual([{ id: "a", score: 1 }]);
-      expect((yield* execution.next("query", cursor))._tag).toBe("None");
-      yield* publishTopicStoreRow(store, { id: "b", score: 2 }, invalidRow);
-      expect((yield* execution.next("query", cursor))._tag).toBe("Some");
-      expect(projectionReads).toBe(1);
-
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
-    }),
-  );
-
-  it.effect("rejects incompatible projection proofs before acquiring raw execution ownership", () =>
-    Effect.gen(function* () {
-      const target = new TopicStore(
-        "projection-ownership-target",
-        Schema.Struct({
-          id: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      const incompatible = new TopicStore(
-        "projection-ownership-incompatible",
-        Schema.Struct({
-          id: Schema.String,
-          score: Schema.String,
-        }),
-        "id",
-        () => {},
-      );
-      const targetReadModel = topicStoreReadModel(target);
-      const targetCompiled = yield* prepareRuntimeRawQuery(
-        "projection-ownership-target",
-        topicStoreRawQueryMetadata(target),
-        { select: ["id"] },
-      );
-      const incompatibleCompiled = yield* prepareRuntimeRawQuery(
-        "projection-ownership-incompatible",
-        topicStoreRawQueryMetadata(incompatible),
-        { select: ["id"] },
-      );
-
-      const emptyRegistryExit = yield* Effect.exit(
-        acquireRawQueryExecution(targetReadModel, incompatibleCompiled),
-      );
-      const emptyRegistryError = Exit.match(emptyRegistryExit, {
-        onFailure: (cause) =>
-          Result.match(Cause.findDefect(cause), {
-            onFailure: () => "missing projection binding error",
-            onSuccess: (defect) =>
-              defect instanceof Error ? defect.message : "unexpected non-error defect",
-          }),
-        onSuccess: () => "unexpected projection binding success",
-      });
-      expect(emptyRegistryError).toBe(
-        "Topic Storage projection schema does not match its compiled proof.",
-      );
-      expect(yield* activeStoreRawQueryExecutionCount(targetReadModel)).toBe(0);
-
-      yield* acquireRawQueryExecution(targetReadModel, targetCompiled);
-      expect(yield* activeStoreRawQueryExecutionCount(targetReadModel)).toBe(1);
-      const existingEntryExit = yield* Effect.exit(
-        acquireRawQueryExecution(targetReadModel, incompatibleCompiled),
-      );
-      const existingEntryError = Exit.match(existingEntryExit, {
-        onFailure: (cause) =>
-          Result.match(Cause.findDefect(cause), {
-            onFailure: () => "missing projection binding error",
-            onSuccess: (defect) =>
-              defect instanceof Error ? defect.message : "unexpected non-error defect",
-          }),
-        onSuccess: () => "unexpected projection binding success",
-      });
-      expect(existingEntryError).toBe(
-        "Topic Storage projection schema does not match its compiled proof.",
-      );
-      expect(yield* activeStoreRawQueryExecutionCount(targetReadModel)).toBe(1);
-
-      yield* releaseRawQueryExecution(targetReadModel, targetCompiled);
-      expect(yield* activeStoreRawQueryExecutionCount(targetReadModel)).toBe(0);
-    }),
-  );
-
-  it.effect("keeps execution caches local to the active query registry", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "registry-isolation",
-        Schema.Struct({
-          id: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      yield* publishTopicStoreRow(store, { id: "a", score: 1 }, invalidRow);
-
-      const compiled = yield* prepareRuntimeRawQuery(
-        "registry-isolation",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          orderBy: [{ field: "score", direction: "desc" }],
-        },
-      );
-      const readModel = topicStoreReadModel(store);
-      const isolatedReadModel = {
-        ...readModel,
-        activeQueries: createActiveQueryRegistry(),
-      };
-
-      yield* acquireRawQueryExecution(readModel, compiled);
-      yield* acquireRawQueryExecution(isolatedReadModel, compiled);
-
-      expect(yield* activeStoreRawQueryExecutionCount(readModel)).toBe(1);
-      expect(yield* activeStoreRawQueryExecutionCount(isolatedReadModel)).toBe(1);
-
-      yield* releaseRawQueryExecution(readModel, compiled);
-      expect(yield* activeStoreRawQueryExecutionCount(readModel)).toBe(0);
-      expect(yield* activeStoreRawQueryExecutionCount(isolatedReadModel)).toBe(1);
-
-      yield* releaseRawQueryExecution(isolatedReadModel, compiled);
-      expect(yield* activeStoreRawQueryExecutionCount(isolatedReadModel)).toBe(0);
-    }),
-  );
-
-  it.effect("keeps materialized execution caches local to the active query registry", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "materialized-registry-isolation",
-        Schema.Struct({
-          id: Schema.String,
-        }),
-        "id",
-        () => {},
-      );
-      const readModel = topicStoreReadModel(store);
-      const isolatedReadModel = {
-        ...readModel,
-        activeQueries: createActiveQueryRegistry(),
-      };
-
-      const emptyEvaluation = (version: number) => ({
-        rows: [],
-        keys: [],
-        window: [],
-        totalRows: 0,
-        version,
-      });
-
-      const emptyDiagnostics = () => ({
-        fullEvaluationCount: 0,
-        patchedEvaluationCount: 0,
-      });
-
-      const firstExecution = yield* acquireMaterializedQueryExecution(
-        readModel,
-        "grouped",
-        emptyResultSemantics,
-        () => ({
-          diagnostics: emptyDiagnostics,
-          incremental: false,
-          latest: () => emptyEvaluation(readModel.version()),
-        }),
-      );
-      const secondExecution = yield* acquireMaterializedQueryExecution(
-        isolatedReadModel,
-        "grouped",
-        emptyResultSemantics,
-        () => ({
-          diagnostics: emptyDiagnostics,
-          incremental: false,
-          latest: () => emptyEvaluation(isolatedReadModel.version()),
-        }),
-      );
-
-      expect(Object.isFrozen(firstExecution)).toBe(true);
-      expect(() => Object.assign(firstExecution, secondExecution)).toThrowError(TypeError);
-      expect(yield* activeStoreRawQueryExecutionCount(readModel)).toBe(1);
-      expect(yield* activeStoreRawQueryExecutionCount(isolatedReadModel)).toBe(1);
-
-      yield* releaseMaterializedQueryExecution(readModel, "grouped");
-      expect(yield* activeStoreRawQueryExecutionCount(readModel)).toBe(0);
-      expect(yield* activeStoreRawQueryExecutionCount(isolatedReadModel)).toBe(1);
-
-      yield* releaseMaterializedQueryExecution(isolatedReadModel, "grouped");
-      expect(yield* activeStoreRawQueryExecutionCount(isolatedReadModel)).toBe(0);
-    }),
-  );
-
-  it.effect("shares base evaluation across different projections", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "projection-sharing",
-        Schema.Struct({
-          id: Schema.String,
-          status: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      yield* publishTopicStoreRow(store, { id: "a", status: "open", score: 1 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "b", status: "open", score: 2 }, invalidRow);
-
-      const idOnly = yield* prepareRuntimeRawQuery(
-        "projection-sharing",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-        },
-      );
-      const idAndScore = yield* prepareRuntimeRawQuery(
-        "projection-sharing",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id", "score"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-        },
-      );
-
-      const idOnlyExecution = yield* acquireRawQueryExecution(topicStoreReadModel(store), idOnly);
-      const idAndScoreExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(store),
-        idAndScore,
-      );
-
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(1);
-      expect(idOnlyExecution.initial("ids").rows).toStrictEqual([{ id: "b" }, { id: "a" }]);
-      expect(idAndScoreExecution.initial("scores").rows).toStrictEqual([
-        { id: "b", score: 2 },
-        { id: "a", score: 1 },
-      ]);
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), idOnly);
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), idAndScore);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(0);
-    }),
-  );
-
-  it.effect("shares base evaluation across different windows", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "window-sharing",
-        Schema.Struct({
-          id: Schema.String,
-          status: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      yield* publishTopicStoreRow(store, { id: "a", status: "open", score: 1 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "b", status: "open", score: 2 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
-
-      const firstWindow = yield* prepareRuntimeRawQuery(
-        "window-sharing",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id", "score"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-          offset: 0,
-          limit: 1,
-        },
-      );
-      const secondWindow = yield* prepareRuntimeRawQuery(
-        "window-sharing",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-          offset: 1,
-          limit: 1,
-        },
-      );
-
-      const firstExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(store),
-        firstWindow,
-      );
-      expect(firstExecution.initial("first").rows).toStrictEqual([{ id: "c", score: 3 }]);
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), secondWindow);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(1);
-      expect(firstExecution.initial("first-after-unknown-release").rows).toStrictEqual([
-        { id: "c", score: 3 },
-      ]);
-
-      const secondExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(store),
-        secondWindow,
-      );
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(1);
-      expect(secondExecution.initial("second").rows).toStrictEqual([{ id: "b" }]);
-
-      const firstCursor = firstExecution.createCursor();
-      const secondCursor = secondExecution.createCursor();
-
-      yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 4 }, invalidRow);
-
-      const firstDelta = yield* firstExecution.next("first", firstCursor);
-      const secondDelta = yield* secondExecution.next("second", secondCursor);
-      expect(Option.getOrThrow(firstDelta)).toStrictEqual({
-        type: "delta",
-        topic: "window-sharing",
-        queryId: "first",
-        fromVersion: 3,
-        toVersion: 4,
-        operations: [
-          {
-            type: "remove",
-            key: "c",
-          },
-          {
-            type: "insert",
-            key: "d",
-            row: {
-              id: "d",
-              score: 4,
-            },
-            index: 0,
-          },
-        ],
-        totalRows: 4,
-      });
-      expect(Option.getOrThrow(secondDelta)).toStrictEqual({
-        type: "delta",
-        topic: "window-sharing",
-        queryId: "second",
-        fromVersion: 3,
-        toVersion: 4,
-        operations: [
-          {
-            type: "remove",
-            key: "b",
-          },
-          {
-            type: "insert",
-            key: "c",
-            row: {
-              id: "c",
-            },
-            index: 0,
-          },
-        ],
-        totalRows: 4,
-      });
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), firstWindow);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(1);
-      expect(secondExecution.initial("second-after-release").rows).toStrictEqual([{ id: "c" }]);
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), secondWindow);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(0);
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), secondWindow);
-    }),
-  );
-
+describe("column-live-view-engine Active Query raw", () => {
   it.effect("updates raw active windows from retained insert-only changes without rescanning", () =>
     Effect.gen(function* () {
       const store = new TopicStore(
@@ -576,18 +34,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-insert-incremental",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -598,7 +56,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
       const cursor = execution.createCursor();
 
@@ -630,8 +88,8 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
-      expect(readModel.changesSince(3)).toBeUndefined();
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
+      expect(queryInterface.changesSince(3)).toBeUndefined();
     }),
   );
 
@@ -652,18 +110,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-insert-incremental-count-only",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -674,7 +132,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
       const cursor = execution.createCursor();
 
@@ -692,49 +150,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
-    }),
-  );
-
-  it.effect("releases retained raw changes when a topic reset clears active queries", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "raw-insert-reset-release",
-        Schema.Struct({
-          id: Schema.String,
-          status: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      yield* publishTopicStoreRow(store, { id: "a", status: "open", score: 1 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "b", status: "open", score: 2 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
-
-      const readModel = topicStoreReadModel(store);
-      const compiled = yield* prepareRuntimeRawQuery(
-        "raw-insert-reset-release",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-          limit: 2,
-        },
-      );
-
-      yield* acquireRawQueryExecution(readModel, compiled);
-      yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 4 }, invalidRow);
-      expect(readModel.changesSince(3)).toBeDefined();
-
-      yield* resetTopicStore(store);
-      expect(yield* activeStoreRawQueryExecutionCount(readModel)).toBe(0);
-
-      yield* publishTopicStoreRow(store, { id: "e", status: "open", score: 5 }, invalidRow);
-      expect(readModel.changesSince(0)).toBeUndefined();
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -755,18 +171,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-match-update-move-up",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -777,7 +193,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
       const cursor = execution.createCursor();
 
@@ -811,7 +227,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -832,18 +248,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-match-update-move-down-fallback",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -854,7 +270,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
       const cursor = execution.createCursor();
 
@@ -886,7 +302,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3, 3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -907,18 +323,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-match-update-unlimited-move-down",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -928,7 +344,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b", "a"]);
       const cursor = execution.createCursor();
 
@@ -962,7 +378,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([undefined]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -985,18 +401,18 @@ describe("column-live-view-engine active query execution", () => {
         yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 1 }, invalidRow);
 
         const scanLimits: Array<number | undefined> = [];
-        const readModel = topicStoreReadModel(store);
-        const observedReadModel = {
-          ...readModel,
-          scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+        const queryInterface = activeQueryTestInterface(store);
+        const observedQueryInterface = {
+          ...queryInterface,
+          scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
             scanLimits.push(plan.limit);
-            return readModel.scanRawWindow(plan);
+            return queryInterface.scanRawWindow(plan);
           },
         };
 
         const compiled = yield* prepareRuntimeRawQuery(
           "raw-match-update-batched-replacements",
-          topicStoreRawQueryMetadata(store),
+          activeQueryTestMetadata(store),
           {
             select: ["id", "score"],
             where: {
@@ -1006,7 +422,7 @@ describe("column-live-view-engine active query execution", () => {
           },
         );
 
-        const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+        const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
         expect(execution.initial("query").keys).toStrictEqual(["a", "b", "c"]);
         const cursor = execution.createCursor();
 
@@ -1056,7 +472,7 @@ describe("column-live-view-engine active query execution", () => {
         });
         expect(scanLimits).toStrictEqual([undefined]);
 
-        yield* releaseRawQueryExecution(observedReadModel, compiled);
+        yield* releaseRawQueryExecution(observedQueryInterface, compiled);
       }),
   );
 
@@ -1078,18 +494,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 4 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-match-update-remove-before-replace",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1100,7 +516,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["d", "c", "b"]);
       const cursor = execution.createCursor();
 
@@ -1148,7 +564,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([4]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -1170,18 +586,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 4 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-match-update-outside-lookahead-fallback",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1192,7 +608,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["d", "c"]);
       const cursor = execution.createCursor();
 
@@ -1224,7 +640,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3, 3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -1246,18 +662,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 7 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-match-update-tail-worsens-fallback",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1268,7 +684,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["a", "b"]);
       const cursor = execution.createCursor();
 
@@ -1303,7 +719,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3, 3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -1324,18 +740,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-pending-insert-match-update",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1346,7 +762,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
       const cursor = execution.createCursor();
 
@@ -1379,7 +795,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -1399,19 +815,19 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "b", status: "open", score: 2 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
         changesSince: () => undefined,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-unavailable-changes-fallback",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1422,7 +838,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["b", "a"]);
       const cursor = execution.createCursor();
 
@@ -1446,7 +862,7 @@ describe("column-live-view-engine active query execution", () => {
       ]);
       expect(scanLimits).toStrictEqual([3, 3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -1466,18 +882,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "b", status: "open", score: 2 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-mixed-insert-incremental",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1488,7 +904,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["b", "a"]);
       const cursor = execution.createCursor();
 
@@ -1519,7 +935,7 @@ describe("column-live-view-engine active query execution", () => {
       ]);
       expect(scanLimits).toStrictEqual([3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -1540,18 +956,18 @@ describe("column-live-view-engine active query execution", () => {
         yield* publishTopicStoreRow(store, { id: "a", status: "open", score: 1 }, invalidRow);
 
         const scanLimits: Array<number | undefined> = [];
-        const readModel = topicStoreReadModel(store);
-        const observedReadModel = {
-          ...readModel,
-          scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+        const queryInterface = activeQueryTestInterface(store);
+        const observedQueryInterface = {
+          ...queryInterface,
+          scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
             scanLimits.push(plan.limit);
-            return readModel.scanRawWindow(plan);
+            return queryInterface.scanRawWindow(plan);
           },
         };
 
         const compiled = yield* prepareRuntimeRawQuery(
           "raw-zero-limit-incremental",
-          topicStoreRawQueryMetadata(store),
+          activeQueryTestMetadata(store),
           {
             select: ["id"],
             where: {
@@ -1561,7 +977,7 @@ describe("column-live-view-engine active query execution", () => {
           },
         );
 
-        const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+        const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
         expect(execution.initial("query").totalRows).toBe(1);
         const cursor = execution.createCursor();
 
@@ -1579,7 +995,7 @@ describe("column-live-view-engine active query execution", () => {
         });
         expect(scanLimits).toStrictEqual([0]);
 
-        yield* releaseRawQueryExecution(observedReadModel, compiled);
+        yield* releaseRawQueryExecution(observedQueryInterface, compiled);
       }),
   );
 
@@ -1609,18 +1025,18 @@ describe("column-live-view-engine active query execution", () => {
       );
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-non-matching-change-incremental",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1645,7 +1061,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       };
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, observedCompiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, observedCompiled);
       expect(execution.initial("query").keys).toStrictEqual(["b", "a"]);
       const cursor = execution.createCursor();
       compareCount = 0;
@@ -1665,7 +1081,7 @@ describe("column-live-view-engine active query execution", () => {
       expect(scanLimits).toStrictEqual([3]);
       expect(compareCount).toBe(0);
 
-      yield* releaseRawQueryExecution(observedReadModel, observedCompiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, observedCompiled);
     }),
   );
 
@@ -1686,7 +1102,7 @@ describe("column-live-view-engine active query execution", () => {
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-retained-insert-slot-sort",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1712,7 +1128,7 @@ describe("column-live-view-engine active query execution", () => {
       };
 
       const execution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(store),
+        activeQueryTestInterface(store),
         observedCompiled,
       );
       expect(execution.initial("query").keys).toStrictEqual(["b", "a"]);
@@ -1726,7 +1142,7 @@ describe("column-live-view-engine active query execution", () => {
       expect(cursor.evaluation.keys).toStrictEqual(["c", "b"]);
       expect(compareCount).toBe(0);
 
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), observedCompiled);
+      yield* releaseRawQueryExecution(activeQueryTestInterface(store), observedCompiled);
     }),
   );
 
@@ -1747,7 +1163,7 @@ describe("column-live-view-engine active query execution", () => {
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-retained-insert-row-sort-fallback",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1771,13 +1187,13 @@ describe("column-live-view-engine active query execution", () => {
           },
         },
       };
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
         compareRawSlots: () => undefined,
       };
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, observedCompiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, observedCompiled);
       expect(execution.initial("query").keys).toStrictEqual(["b", "a"]);
       const cursor = execution.createCursor();
       compareCount = 0;
@@ -1789,7 +1205,7 @@ describe("column-live-view-engine active query execution", () => {
       expect(cursor.evaluation.keys).toStrictEqual(["c", "b"]);
       expect(compareCount).toBeGreaterThan(0);
 
-      yield* releaseRawQueryExecution(observedReadModel, observedCompiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, observedCompiled);
     }),
   );
 
@@ -1810,7 +1226,7 @@ describe("column-live-view-engine active query execution", () => {
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-retained-insert-missing-slot-fallback",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1834,13 +1250,13 @@ describe("column-live-view-engine active query execution", () => {
           },
         },
       };
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        slotForKey: (key: string) => (key === "c" ? undefined : readModel.slotForKey?.(key)),
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        slotForKey: (key: string) => (key === "c" ? undefined : queryInterface.slotForKey?.(key)),
       };
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, observedCompiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, observedCompiled);
       expect(execution.initial("query").keys).toStrictEqual(["b", "a"]);
       const cursor = execution.createCursor();
       compareCount = 0;
@@ -1852,7 +1268,7 @@ describe("column-live-view-engine active query execution", () => {
       expect(cursor.evaluation.keys).toStrictEqual(["c", "b"]);
       expect(compareCount).toBeGreaterThan(0);
 
-      yield* releaseRawQueryExecution(observedReadModel, observedCompiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, observedCompiled);
     }),
   );
 
@@ -1878,10 +1294,10 @@ describe("column-live-view-engine active query execution", () => {
           invalidRow,
         );
 
-        const readModel = topicStoreReadModel(store);
+        const queryInterface = activeQueryTestInterface(store);
         const compiled = yield* prepareRuntimeRawQuery(
           "raw-noop-then-visible-version",
-          topicStoreRawQueryMetadata(store),
+          activeQueryTestMetadata(store),
           {
             select: ["id", "score"],
             where: {
@@ -1892,7 +1308,7 @@ describe("column-live-view-engine active query execution", () => {
           },
         );
 
-        const execution = yield* acquireRawQueryExecution(readModel, compiled);
+        const execution = yield* acquireRawQueryExecution(queryInterface, compiled);
         expect(execution.initial("query").keys).toStrictEqual(["b", "a"]);
         const cursor = execution.createCursor();
 
@@ -1935,7 +1351,7 @@ describe("column-live-view-engine active query execution", () => {
           totalRows: 3,
         });
 
-        yield* releaseRawQueryExecution(readModel, compiled);
+        yield* releaseRawQueryExecution(queryInterface, compiled);
       }),
   );
 
@@ -1957,18 +1373,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "d", status: "closed", score: 4 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-retained-update-enters-predicate",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -1979,7 +1395,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
       const cursor = execution.createCursor();
 
@@ -2011,7 +1427,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -2032,18 +1448,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-same-key-pending-insert-removed",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -2054,7 +1470,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
       const cursor = execution.createCursor();
 
@@ -2065,7 +1481,7 @@ describe("column-live-view-engine active query execution", () => {
       expect(Option.isNone(delta)).toBe(true);
       expect(scanLimits).toStrictEqual([3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -2087,18 +1503,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 70 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-retained-removal-consumes-lookahead",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -2109,7 +1525,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["a", "b"]);
       const cursor = execution.createCursor();
 
@@ -2170,7 +1586,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3, 3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -2194,18 +1610,18 @@ describe("column-live-view-engine active query execution", () => {
         yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 4 }, invalidRow);
 
         const scanLimits: Array<number | undefined> = [];
-        const readModel = topicStoreReadModel(store);
-        const observedReadModel = {
-          ...readModel,
-          scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+        const queryInterface = activeQueryTestInterface(store);
+        const observedQueryInterface = {
+          ...queryInterface,
+          scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
             scanLimits.push(plan.limit);
-            return readModel.scanRawWindow(plan);
+            return queryInterface.scanRawWindow(plan);
           },
         };
 
         const compiled = yield* prepareRuntimeRawQuery(
           "raw-retained-exhausted-lookahead-rejects-later-insert",
-          topicStoreRawQueryMetadata(store),
+          activeQueryTestMetadata(store),
           {
             select: ["id", "score"],
             where: {
@@ -2216,7 +1632,7 @@ describe("column-live-view-engine active query execution", () => {
           },
         );
 
-        const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+        const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
         expect(execution.initial("query").keys).toStrictEqual(["d", "c"]);
         const cursor = execution.createCursor();
 
@@ -2290,7 +1706,7 @@ describe("column-live-view-engine active query execution", () => {
         });
         expect(scanLimits).toStrictEqual([3, 3]);
 
-        yield* releaseRawQueryExecution(observedReadModel, compiled);
+        yield* releaseRawQueryExecution(observedQueryInterface, compiled);
       }),
   );
 
@@ -2314,18 +1730,18 @@ describe("column-live-view-engine active query execution", () => {
         yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
         const scanLimits: Array<number | undefined> = [];
-        const readModel = topicStoreReadModel(store);
-        const observedReadModel = {
-          ...readModel,
-          scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+        const queryInterface = activeQueryTestInterface(store);
+        const observedQueryInterface = {
+          ...queryInterface,
+          scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
             scanLimits.push(plan.limit);
-            return readModel.scanRawWindow(plan);
+            return queryInterface.scanRawWindow(plan);
           },
         };
 
         const compiled = yield* prepareRuntimeRawQuery(
           "raw-delete-outside-retained-window",
-          topicStoreRawQueryMetadata(store),
+          activeQueryTestMetadata(store),
           {
             select: ["id", "score"],
             where: {
@@ -2336,7 +1752,7 @@ describe("column-live-view-engine active query execution", () => {
           },
         );
 
-        const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+        const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
         expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
         const cursor = execution.createCursor();
 
@@ -2354,7 +1770,7 @@ describe("column-live-view-engine active query execution", () => {
         });
         expect(scanLimits).toStrictEqual([3]);
 
-        yield* releaseRawQueryExecution(observedReadModel, compiled);
+        yield* releaseRawQueryExecution(observedQueryInterface, compiled);
       }),
   );
 
@@ -2378,18 +1794,18 @@ describe("column-live-view-engine active query execution", () => {
         yield* publishTopicStoreRow(store, { id: "d", status: "closed", score: 4 }, invalidRow);
 
         const scanLimits: Array<number | undefined> = [];
-        const readModel = topicStoreReadModel(store);
-        const observedReadModel = {
-          ...readModel,
-          scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+        const queryInterface = activeQueryTestInterface(store);
+        const observedQueryInterface = {
+          ...queryInterface,
+          scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
             scanLimits.push(plan.limit);
-            return readModel.scanRawWindow(plan);
+            return queryInterface.scanRawWindow(plan);
           },
         };
 
         const compiled = yield* prepareRuntimeRawQuery(
           "raw-zero-limit-mixed-incremental",
-          topicStoreRawQueryMetadata(store),
+          activeQueryTestMetadata(store),
           {
             select: ["id"],
             where: {
@@ -2399,7 +1815,7 @@ describe("column-live-view-engine active query execution", () => {
           },
         );
 
-        const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+        const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
         expect(execution.initial("query").totalRows).toBe(2);
         const cursor = execution.createCursor();
 
@@ -2427,7 +1843,7 @@ describe("column-live-view-engine active query execution", () => {
         });
         expect(scanLimits).toStrictEqual([0]);
 
-        yield* releaseRawQueryExecution(observedReadModel, compiled);
+        yield* releaseRawQueryExecution(observedQueryInterface, compiled);
       }),
   );
 
@@ -2448,18 +1864,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-visible-delete-fallback",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -2470,7 +1886,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["c", "b"]);
       const cursor = execution.createCursor();
 
@@ -2502,7 +1918,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -2524,18 +1940,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 4 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const compiled = yield* prepareRuntimeRawQuery(
         "raw-visible-delete-exhausted-lookahead",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -2546,7 +1962,7 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const execution = yield* acquireRawQueryExecution(observedReadModel, compiled);
+      const execution = yield* acquireRawQueryExecution(observedQueryInterface, compiled);
       expect(execution.initial("query").keys).toStrictEqual(["d", "c"]);
       const cursor = execution.createCursor();
 
@@ -2592,7 +2008,7 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([3, 3]);
 
-      yield* releaseRawQueryExecution(observedReadModel, compiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, compiled);
     }),
   );
 
@@ -2616,18 +2032,18 @@ describe("column-live-view-engine active query execution", () => {
       yield* publishTopicStoreRow(store, { id: "f", status: "open", score: 6 }, invalidRow);
 
       const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+      const queryInterface = activeQueryTestInterface(store);
+      const observedQueryInterface = {
+        ...queryInterface,
+        scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
           scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
+          return queryInterface.scanRawWindow(plan);
         },
       };
 
       const wideCompiled = yield* prepareRuntimeRawQuery(
         "raw-visible-delete-shared-wide-lookahead",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -2639,7 +2055,7 @@ describe("column-live-view-engine active query execution", () => {
       );
       const narrowCompiled = yield* prepareRuntimeRawQuery(
         "raw-visible-delete-shared-wide-lookahead",
-        topicStoreRawQueryMetadata(store),
+        activeQueryTestMetadata(store),
         {
           select: ["id", "score"],
           where: {
@@ -2650,8 +2066,11 @@ describe("column-live-view-engine active query execution", () => {
         },
       );
 
-      const wideExecution = yield* acquireRawQueryExecution(observedReadModel, wideCompiled);
-      const narrowExecution = yield* acquireRawQueryExecution(observedReadModel, narrowCompiled);
+      const wideExecution = yield* acquireRawQueryExecution(observedQueryInterface, wideCompiled);
+      const narrowExecution = yield* acquireRawQueryExecution(
+        observedQueryInterface,
+        narrowCompiled,
+      );
       expect(wideExecution.initial("wide-query").keys).toStrictEqual([
         "f",
         "e",
@@ -2705,8 +2124,8 @@ describe("column-live-view-engine active query execution", () => {
       });
       expect(scanLimits).toStrictEqual([192]);
 
-      yield* releaseRawQueryExecution(observedReadModel, narrowCompiled);
-      yield* releaseRawQueryExecution(observedReadModel, wideCompiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, narrowCompiled);
+      yield* releaseRawQueryExecution(observedQueryInterface, wideCompiled);
     }),
   );
 
@@ -2735,18 +2154,18 @@ describe("column-live-view-engine active query execution", () => {
         );
 
         const scanLimits: Array<number | undefined> = [];
-        const readModel = topicStoreReadModel(store);
-        const observedReadModel = {
-          ...readModel,
-          scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
+        const queryInterface = activeQueryTestInterface(store);
+        const observedQueryInterface = {
+          ...queryInterface,
+          scanRawWindow: (plan: Parameters<typeof queryInterface.scanRawWindow>[0]) => {
             scanLimits.push(plan.limit);
-            return readModel.scanRawWindow(plan);
+            return queryInterface.scanRawWindow(plan);
           },
         };
 
         const wideCompiled = yield* prepareRuntimeRawQuery(
           "raw-visible-delete-insert-preserves-lookahead",
-          topicStoreRawQueryMetadata(store),
+          activeQueryTestMetadata(store),
           {
             select: ["id", "score"],
             where: {
@@ -2758,7 +2177,7 @@ describe("column-live-view-engine active query execution", () => {
         );
         const narrowCompiled = yield* prepareRuntimeRawQuery(
           "raw-visible-delete-insert-preserves-lookahead",
-          topicStoreRawQueryMetadata(store),
+          activeQueryTestMetadata(store),
           {
             select: ["id", "score"],
             where: {
@@ -2769,8 +2188,11 @@ describe("column-live-view-engine active query execution", () => {
           },
         );
 
-        const wideExecution = yield* acquireRawQueryExecution(observedReadModel, wideCompiled);
-        const narrowExecution = yield* acquireRawQueryExecution(observedReadModel, narrowCompiled);
+        const wideExecution = yield* acquireRawQueryExecution(observedQueryInterface, wideCompiled);
+        const narrowExecution = yield* acquireRawQueryExecution(
+          observedQueryInterface,
+          narrowCompiled,
+        );
         expect(wideExecution.initial("wide-query").keys.slice(0, 3)).toStrictEqual([
           "row-219",
           "row-218",
@@ -2881,682 +2303,8 @@ describe("column-live-view-engine active query execution", () => {
         });
         expect(scanLimits).toStrictEqual([192]);
 
-        yield* releaseRawQueryExecution(observedReadModel, narrowCompiled);
-        yield* releaseRawQueryExecution(observedReadModel, wideCompiled);
+        yield* releaseRawQueryExecution(observedQueryInterface, narrowCompiled);
+        yield* releaseRawQueryExecution(observedQueryInterface, wideCompiled);
       }),
-  );
-
-  it.effect("shrinks shared base windows immediately when larger windows release", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "window-shrink",
-        Schema.Struct({
-          id: Schema.String,
-          status: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      yield* publishTopicStoreRow(store, { id: "a", status: "open", score: 1 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "b", status: "open", score: 2 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "d", status: "open", score: 4 }, invalidRow);
-
-      const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
-          scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
-        },
-      };
-
-      const wideWindow = yield* prepareRuntimeRawQuery(
-        "window-shrink",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-          offset: 0,
-          limit: 3,
-        },
-      );
-      const narrowWindow = yield* prepareRuntimeRawQuery(
-        "window-shrink",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-          offset: 1,
-          limit: 1,
-        },
-      );
-
-      const wideExecution = yield* acquireRawQueryExecution(observedReadModel, wideWindow);
-      expect(wideExecution.initial("wide").keys).toStrictEqual(["d", "c", "b"]);
-
-      const narrowExecution = yield* acquireRawQueryExecution(observedReadModel, narrowWindow);
-      expect(narrowExecution.initial("narrow").keys).toStrictEqual(["c"]);
-
-      yield* releaseRawQueryExecution(observedReadModel, wideWindow);
-      expect(scanLimits).toStrictEqual([4, 3]);
-      expect(narrowExecution.initial("narrow-after-shrink").keys).toStrictEqual(["c"]);
-
-      yield* releaseRawQueryExecution(observedReadModel, narrowWindow);
-    }),
-  );
-
-  it.effect("compacts unbounded shared base windows when unbounded windows release", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "window-unbounded",
-        Schema.Struct({
-          id: Schema.String,
-          status: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      yield* publishTopicStoreRow(store, { id: "a", status: "open", score: 1 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "b", status: "open", score: 2 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "c", status: "open", score: 3 }, invalidRow);
-
-      const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
-          scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
-        },
-      };
-
-      const boundedWindow = yield* prepareRuntimeRawQuery(
-        "window-unbounded",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-          offset: 1,
-          limit: 1,
-        },
-      );
-      const unboundedWindow = yield* prepareRuntimeRawQuery(
-        "window-unbounded",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-        },
-      );
-
-      const boundedExecution = yield* acquireRawQueryExecution(observedReadModel, boundedWindow);
-      expect(boundedExecution.initial("bounded").keys).toStrictEqual(["b"]);
-
-      const unboundedExecution = yield* acquireRawQueryExecution(
-        observedReadModel,
-        unboundedWindow,
-      );
-      expect(unboundedExecution.initial("unbounded").keys).toStrictEqual(["c", "b", "a"]);
-
-      yield* releaseRawQueryExecution(observedReadModel, unboundedWindow);
-      expect(scanLimits).toStrictEqual([3, undefined, 3]);
-      expect(boundedExecution.initial("bounded-after-compact").keys).toStrictEqual(["b"]);
-
-      yield* releaseRawQueryExecution(observedReadModel, boundedWindow);
-    }),
-  );
-
-  it.effect("does not expand shared base rows for zero-limit windows", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "window-zero-limit",
-        Schema.Struct({
-          id: Schema.String,
-          status: Schema.String,
-          score: Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      yield* publishTopicStoreRow(store, { id: "a", status: "open", score: 1 }, invalidRow);
-      yield* publishTopicStoreRow(store, { id: "b", status: "open", score: 2 }, invalidRow);
-
-      const scanLimits: Array<number | undefined> = [];
-      const readModel = topicStoreReadModel(store);
-      const observedReadModel = {
-        ...readModel,
-        scanRawWindow: (plan: Parameters<typeof readModel.scanRawWindow>[0]) => {
-          scanLimits.push(plan.limit);
-          return readModel.scanRawWindow(plan);
-        },
-      };
-
-      const zeroWindow = yield* prepareRuntimeRawQuery(
-        "window-zero-limit",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          where: {
-            status: "open",
-          },
-          orderBy: [{ field: "score", direction: "desc" }],
-          offset: 10,
-          limit: 0,
-        },
-      );
-
-      const zeroExecution = yield* acquireRawQueryExecution(observedReadModel, zeroWindow);
-      const initial = zeroExecution.initial("zero");
-      expect(scanLimits).toStrictEqual([0]);
-      expect(initial.rows).toStrictEqual([]);
-      expect(initial.keys).toStrictEqual([]);
-      expect(initial.totalRows).toBe(2);
-
-      yield* releaseRawQueryExecution(observedReadModel, zeroWindow);
-    }),
-  );
-
-  it.effect("no-ops release when execution cache does not contain a query", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "numbers",
-        Schema.Struct({ id: Schema.String, score: Schema.Number }),
-        "id",
-        () => {},
-      );
-      const compiled = yield* prepareRuntimeRawQuery("numbers", topicStoreRawQueryMetadata(store), {
-        select: ["id"],
-        where: {
-          score: 1,
-        },
-      });
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), compiled);
-    }),
-  );
-
-  it.effect("covers query cache key paths for numeric, bigint, and BigDecimal filters", () =>
-    Effect.gen(function* () {
-      const numericStore = new TopicStore(
-        "numbers",
-        Schema.Struct({ id: Schema.String, score: Schema.Number }),
-        "id",
-        () => {},
-      );
-      const bigintStore = new TopicStore(
-        "bigints",
-        Schema.Struct({ id: Schema.String, amount: Schema.BigInt }),
-        "id",
-        () => {},
-      );
-      const decimalStore = new TopicStore(
-        "decimals",
-        Schema.Struct({ id: Schema.String, price: Schema.BigDecimal }),
-        "id",
-        () => {},
-      );
-
-      yield* publishTopicStoreRow(numericStore, { id: "a", score: 10 }, invalidRow);
-      yield* publishTopicStoreRow(bigintStore, { id: "b", amount: 5n }, invalidRow);
-      yield* publishTopicStoreRow(
-        decimalStore,
-        { id: "c", price: fromStringUnsafe("1.23") },
-        invalidRow,
-      );
-
-      const infFilter = yield* prepareRuntimeRawQuery(
-        "numbers",
-        topicStoreRawQueryMetadata(numericStore),
-        {
-          select: ["id", "score"],
-          where: {
-            score: Number.POSITIVE_INFINITY,
-          },
-        },
-      );
-      const nanFilter = yield* prepareRuntimeRawQuery(
-        "numbers",
-        topicStoreRawQueryMetadata(numericStore),
-        {
-          select: ["id", "score"],
-          where: {
-            score: Number.NaN,
-          },
-        },
-      );
-      const zeroFilter = yield* prepareRuntimeRawQuery(
-        "numbers",
-        topicStoreRawQueryMetadata(numericStore),
-        {
-          select: ["id", "score"],
-          where: {
-            score: 10,
-          },
-        },
-      );
-      const positiveZeroFilter = yield* prepareRuntimeRawQuery(
-        "numbers",
-        topicStoreRawQueryMetadata(numericStore),
-        {
-          select: ["id", "score"],
-          where: {
-            score: 0,
-          },
-        },
-      );
-      const negativeZeroFilter = yield* prepareRuntimeRawQuery(
-        "numbers",
-        topicStoreRawQueryMetadata(numericStore),
-        {
-          select: ["id", "score"],
-          where: {
-            score: -0,
-          },
-        },
-      );
-      const offsetFilter = yield* prepareRuntimeRawQuery(
-        "numbers",
-        topicStoreRawQueryMetadata(numericStore),
-        {
-          select: ["id", "score"],
-          offset: 1,
-        },
-      );
-      const bigIntFilter = yield* prepareRuntimeRawQuery(
-        "bigints",
-        topicStoreRawQueryMetadata(bigintStore),
-        {
-          select: ["id", "amount"],
-          where: {
-            amount: 3n,
-          },
-        },
-      );
-      const decimalFilter = yield* prepareRuntimeRawQuery(
-        "decimals",
-        topicStoreRawQueryMetadata(decimalStore),
-        {
-          select: ["id", "price"],
-          where: {
-            price: fromStringUnsafe("1.23"),
-          },
-        },
-      );
-
-      const infExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(numericStore),
-        infFilter,
-      );
-      const nanExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(numericStore),
-        nanFilter,
-      );
-      const zeroExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(numericStore),
-        zeroFilter,
-      );
-      yield* acquireRawQueryExecution(topicStoreReadModel(numericStore), positiveZeroFilter);
-      yield* acquireRawQueryExecution(topicStoreReadModel(numericStore), negativeZeroFilter);
-      const offsetExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(numericStore),
-        offsetFilter,
-      );
-      const bigintExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(bigintStore),
-        bigIntFilter,
-      );
-      const decimalExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(decimalStore),
-        decimalFilter,
-      );
-
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(numericStore))).toBe(5);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(bigintStore))).toBe(1);
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(decimalStore))).toBe(1);
-
-      const infCursor = infExecution.createCursor();
-      const nanCursor = nanExecution.createCursor();
-      const zeroCursor = zeroExecution.createCursor();
-      const offsetCursor = offsetExecution.createCursor();
-      const bigintCursor = bigintExecution.createCursor();
-      const decimalCursor = decimalExecution.createCursor();
-
-      expect(infExecution.initial("q").totalRows).toBe(0);
-      expect(nanExecution.initial("q").totalRows).toBe(0);
-      expect(zeroExecution.initial("q").totalRows).toBe(1);
-      expect(offsetExecution.initial("q").totalRows).toBe(1);
-      expect(bigintExecution.initial("q").totalRows).toBe(0);
-      expect(decimalExecution.initial("q").totalRows).toBe(1);
-
-      expect((yield* infExecution.next("q", infCursor))._tag).toBe("None");
-      expect((yield* nanExecution.next("q", nanCursor))._tag).toBe("None");
-      expect((yield* zeroExecution.next("q", zeroCursor))._tag).toBe("None");
-      expect((yield* offsetExecution.next("q", offsetCursor))._tag).toBe("None");
-      expect((yield* bigintExecution.next("q", bigintCursor))._tag).toBe("None");
-      expect((yield* decimalExecution.next("q", decimalCursor))._tag).toBe("None");
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(numericStore), infFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(numericStore), nanFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(numericStore), zeroFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(numericStore), positiveZeroFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(numericStore), negativeZeroFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(numericStore), offsetFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(bigintStore), bigIntFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(decimalStore), decimalFilter);
-    }),
-  );
-
-  it.effect("covers cache keys for nullish, array, object, and boolean filter values", () =>
-    Effect.gen(function* () {
-      const eventStore = new TopicStore(
-        "events",
-        Schema.Struct({
-          id: Schema.String,
-          label: Schema.String,
-          tags: Schema.Array(Schema.String),
-          metadata: Schema.Struct({
-            kind: Schema.String,
-            scope: Schema.String,
-          }),
-          active: Schema.Boolean,
-        }),
-        "id",
-        () => {},
-      );
-
-      yield* publishTopicStoreRow(
-        eventStore,
-        {
-          id: "a",
-          label: "foo",
-          tags: ["open", "closed"],
-          metadata: { kind: "test", scope: "global" },
-          active: true,
-        },
-        invalidRow,
-      );
-
-      const undefinedFilter = yield* prepareRuntimeRawQuery(
-        "events",
-        topicStoreRawQueryMetadata(eventStore),
-        {
-          select: ["id", "label"],
-          where: {
-            label: undefined,
-          },
-        },
-      );
-      const nullFilter = yield* prepareRuntimeRawQuery(
-        "events",
-        topicStoreRawQueryMetadata(eventStore),
-        {
-          select: ["id", "label"],
-          where: {
-            label: null,
-          },
-        },
-      );
-      const arrayFilter = yield* prepareRuntimeRawQuery(
-        "events",
-        topicStoreRawQueryMetadata(eventStore),
-        {
-          select: ["id", "tags"],
-          where: {
-            tags: ["open", "closed"],
-          },
-        },
-      );
-      const objectFilter = yield* prepareRuntimeRawQuery(
-        "events",
-        topicStoreRawQueryMetadata(eventStore),
-        {
-          select: ["id", "metadata"],
-          where: {
-            metadata: { kind: "test", scope: "global" },
-          },
-        },
-      );
-      const booleanFilter = yield* prepareRuntimeRawQuery(
-        "events",
-        topicStoreRawQueryMetadata(eventStore),
-        {
-          select: ["id", "active"],
-          where: {
-            active: true,
-          },
-        },
-      );
-
-      const undefinedExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(eventStore),
-        undefinedFilter,
-      );
-      const nullExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(eventStore),
-        nullFilter,
-      );
-      const arrayExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(eventStore),
-        arrayFilter,
-      );
-      const objectExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(eventStore),
-        objectFilter,
-      );
-      const booleanExecution = yield* acquireRawQueryExecution(
-        topicStoreReadModel(eventStore),
-        booleanFilter,
-      );
-
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(eventStore))).toBe(5);
-
-      expect(undefinedExecution.initial("query").totalRows).toBe(0);
-      expect(nullExecution.initial("query").totalRows).toBe(0);
-      expect(arrayExecution.initial("query").totalRows).toBe(1);
-      expect(objectExecution.initial("query").totalRows).toBe(1);
-      expect(booleanExecution.initial("query").totalRows).toBe(1);
-
-      const undefinedCursor = undefinedExecution.createCursor();
-      const nullCursor = nullExecution.createCursor();
-      const arrayCursor = arrayExecution.createCursor();
-      const objectCursor = objectExecution.createCursor();
-      const booleanCursor = booleanExecution.createCursor();
-
-      expect((yield* undefinedExecution.next("query", undefinedCursor))._tag).toBe("None");
-      expect((yield* nullExecution.next("query", nullCursor))._tag).toBe("None");
-      expect((yield* arrayExecution.next("query", arrayCursor))._tag).toBe("None");
-      expect((yield* objectExecution.next("query", objectCursor))._tag).toBe("None");
-      expect((yield* booleanExecution.next("query", booleanCursor))._tag).toBe("None");
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(eventStore), undefinedFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(eventStore), nullFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(eventStore), arrayFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(eventStore), objectFilter);
-      yield* releaseRawQueryExecution(topicStoreReadModel(eventStore), booleanFilter);
-    }),
-  );
-
-  it.effect("does not collide delimiter-bearing field names in cache keys", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "delimiter-fields",
-        Schema.Struct({
-          id: Schema.String,
-          a: Schema.Number,
-          b: Schema.Number,
-          "a:asc;b": Schema.Number,
-        }),
-        "id",
-        () => {},
-      );
-      const splitFieldsQuery = yield* prepareRuntimeRawQuery(
-        "delimiter-fields",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          orderBy: [
-            {
-              field: "a",
-              direction: "asc",
-            },
-            {
-              field: "b",
-              direction: "desc",
-            },
-          ],
-        },
-      );
-      const delimiterFieldQuery = yield* prepareRuntimeRawQuery(
-        "delimiter-fields",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id"],
-          orderBy: [
-            {
-              field: "a:asc;b",
-              direction: "desc",
-            },
-          ],
-        },
-      );
-
-      yield* acquireRawQueryExecution(topicStoreReadModel(store), splitFieldsQuery);
-      yield* acquireRawQueryExecution(topicStoreReadModel(store), delimiterFieldQuery);
-
-      expect(yield* activeStoreRawQueryExecutionCount(topicStoreReadModel(store))).toBe(2);
-
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), splitFieldsQuery);
-      yield* releaseRawQueryExecution(topicStoreReadModel(store), delimiterFieldQuery);
-    }),
-  );
-
-  it.effect("canonicalizes schema-backed null-prototype filter values before cache-keying", () =>
-    Effect.gen(function* () {
-      const store = new TopicStore(
-        "special",
-        Schema.Struct({
-          id: Schema.String,
-          payload: Schema.Record(Schema.String, Schema.String),
-        }),
-        "id",
-        () => {},
-      );
-      const queryPayload: Record<string, string> = Object.create(null);
-      queryPayload["venue"] = "xnys";
-
-      const nullPrototypeQuery = yield* prepareRuntimeRawQuery(
-        "special",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id", "payload"],
-          where: {
-            payload: queryPayload,
-          },
-        },
-      );
-      const plainRecordQuery = yield* prepareRuntimeRawQuery(
-        "special",
-        topicStoreRawQueryMetadata(store),
-        {
-          select: ["id", "payload"],
-          where: {
-            payload: { venue: "xnys" },
-          },
-        },
-      );
-      const nullPrototypeGrouped = yield* prepareRuntimeGroupedQuery(
-        "special",
-        topicStoreRawQueryMetadata(store),
-        {
-          groupBy: ["payload"],
-          aggregates: { rowCount: { aggFunc: "count" } },
-          where: { payload: queryPayload },
-        },
-      );
-      const plainRecordGrouped = yield* prepareRuntimeGroupedQuery(
-        "special",
-        topicStoreRawQueryMetadata(store),
-        {
-          groupBy: ["payload"],
-          aggregates: { rowCount: { aggFunc: "count" } },
-          where: { payload: { venue: "xnys" } },
-        },
-      );
-
-      expect(nullPrototypeQuery.plan.queryCacheKey).toBe(plainRecordQuery.plan.queryCacheKey);
-      expect(nullPrototypeGrouped.cacheKey).toBe(plainRecordGrouped.cacheKey);
-    }),
-  );
-
-  it.effect("rejects non-serializable filter values before cache-keying", () =>
-    Effect.gen(function* () {
-      const firstFunction = () => "first";
-      const firstSymbol = Symbol("first");
-      const firstMap = new Map([["marker", "first"]]);
-      const store = new TopicStore(
-        "special-non-serializable",
-        Schema.Struct({
-          id: Schema.String,
-          marker: Schema.Unknown,
-        }),
-        "id",
-        () => {},
-      );
-
-      const functionFilter = yield* Effect.flip(
-        prepareRuntimeRawQuery("special-non-serializable", topicStoreRawQueryMetadata(store), {
-          select: ["id", "marker"],
-          where: {
-            marker: firstFunction,
-          },
-        }),
-      );
-      expect(functionFilter).toMatchObject({
-        _tag: "InvalidQueryError",
-        message: expect.stringContaining("unsupported query value"),
-      });
-
-      const symbolFilter = yield* Effect.flip(
-        prepareRuntimeRawQuery("special-non-serializable", topicStoreRawQueryMetadata(store), {
-          select: ["id", "marker"],
-          where: {
-            marker: firstSymbol,
-          },
-        }),
-      );
-      expect(symbolFilter).toMatchObject({
-        _tag: "InvalidQueryError",
-        message: expect.stringContaining("unsupported query value"),
-      });
-
-      const mapFilter = yield* Effect.flip(
-        prepareRuntimeRawQuery("special-non-serializable", topicStoreRawQueryMetadata(store), {
-          select: ["id", "marker"],
-          where: {
-            marker: firstMap,
-          },
-        }),
-      );
-      expect(mapFilter).toMatchObject({
-        _tag: "InvalidQueryError",
-        message: expect.stringContaining("unsupported query value"),
-      });
-    }),
   );
 });
