+++ packages/column-live-view-engine/src/active-query.ts
@@ -1,19 +1,12 @@
-import { Effect, Option } from "effect";
-import type { DeltaEvent, SnapshotEvent } from "@effect-view-server/config";
-import type { QueryEvaluation } from "./query-result";
-import type { TopicRawWindowScan } from "./raw-window-scan";
-import type { TopicRowScan } from "./row-scan";
-import {
-  activeRawQueryExecutionCount,
-  clearRawQueryExecutions,
-  type RawQueryExecutionSlot,
-} from "./active-raw-query";
-import type { MaterializedQueryExecutionSlot } from "./active-materialized-query";
+import { Effect } from "effect";
+import { activeRawQueryExecutionCount, clearRawQueryExecutions } from "./active-raw-query";
 import {
   activeMaterializedQueryExecutionCount,
   activeMaterializedQueryExecutionModeCounts,
   clearMaterializedQueryExecutions,
 } from "./active-materialized-query";
+import type { ActiveQueryExecutionCounts, ActiveQueryRegistry } from "./active-query-contract";
+
 export {
   acquireRawQueryExecution,
   evaluateRawQuery,
@@ -24,84 +17,43 @@ export {
   acquireMaterializedQueryExecution,
   releaseMaterializedQueryExecution,
 } from "./active-materialized-query";
-export type { MaterializedQueryExecution } from "./active-materialized-query";
-
-type RowObject = object;
-
-export type ActiveQueryStoreState = TopicRowScan<object> &
-  TopicRawWindowScan<object> & {
-    readonly activeQueries: ActiveQueryRegistry;
-    readonly releaseChanges: () => void;
-    readonly retainChanges: () => void;
-    readonly topic: string;
-  };
-
-export type LiveQueryExecutionCursor = {
-  evaluation: QueryEvaluation<RowObject>;
-};
-
-type RawQueryExecutionUpdate<ResultRow extends RowObject> = Effect.Effect<
-  Option.Option<DeltaEvent<ResultRow>>,
-  never,
-  never
->;
-
-export type LiveQueryExecution<ResultRow extends RowObject> = {
-  readonly initial: (queryId: string) => SnapshotEvent<ResultRow>;
-  readonly createCursor: () => LiveQueryExecutionCursor;
-  readonly next: (
-    queryId: string,
-    cursor: LiveQueryExecutionCursor,
-  ) => RawQueryExecutionUpdate<ResultRow>;
-};
-
-export type RawQueryExecution<ResultRow extends RowObject> = LiveQueryExecution<ResultRow>;
-
-export type ActiveQueryRegistry = {
-  readonly raw: Map<string, RawQueryExecutionSlot>;
-  readonly materialized: Map<string, MaterializedQueryExecutionSlot>;
-};
-
-export type ActiveQueryExecutionCounts = {
-  readonly activeFallbackGroupedViews: number;
-  readonly activeIncrementalGroupedViews: number;
-  readonly activeViews: number;
-  readonly groupedFullEvaluationCount: number;
-  readonly groupedPatchedEvaluationCount: number;
-};
-
-export const createActiveQueryRegistry = (): ActiveQueryRegistry => ({
-  raw: new Map(),
-  materialized: new Map(),
-});
+export {
+  createActiveQueryRegistry,
+  type ActiveQueryExecutionCounts,
+  type ActiveQueryRegistry,
+  type LiveQueryExecution,
+  type LiveQueryExecutionCursor,
+  type MaterializedQueryExecution,
+  type RawQueryExecution,
+} from "./active-query-contract";
 
 export const clearStoreRawQueryExecutions = Effect.fn(
   "ColumnLiveViewEngine.activeQuery.clearStore",
-)((store: ActiveQueryStoreState) =>
+)((registry: ActiveQueryRegistry) =>
   Effect.uninterruptible(
     Effect.gen(function* () {
-      yield* clearRawQueryExecutions(store);
-      yield* clearMaterializedQueryExecutions(store);
+      yield* clearRawQueryExecutions(registry);
+      yield* clearMaterializedQueryExecutions(registry);
     }),
   ),
 );
 
 export const activeStoreRawQueryExecutionCount = Effect.fn(
   "ColumnLiveViewEngine.activeQuery.countStore",
-)((store: ActiveQueryStoreState) =>
+)((registry: ActiveQueryRegistry) =>
   Effect.gen(function* () {
-    const rawCount = yield* activeRawQueryExecutionCount(store);
-    const materializedCount = yield* activeMaterializedQueryExecutionCount(store);
+    const rawCount = yield* activeRawQueryExecutionCount(registry);
+    const materializedCount = yield* activeMaterializedQueryExecutionCount(registry);
     return rawCount + materializedCount;
   }),
 );
 
 export const activeStoreQueryExecutionCounts = Effect.fn(
   "ColumnLiveViewEngine.activeQuery.countStoreModes",
-)((store: ActiveQueryStoreState) =>
+)((registry: ActiveQueryRegistry) =>
   Effect.gen(function* () {
-    const rawCount = yield* activeRawQueryExecutionCount(store);
-    const materializedCounts = yield* activeMaterializedQueryExecutionModeCounts(store);
+    const rawCount = yield* activeRawQueryExecutionCount(registry);
+    const materializedCounts = yield* activeMaterializedQueryExecutionModeCounts(registry);
     return {
       activeFallbackGroupedViews: materializedCounts.activeFallback,
       activeIncrementalGroupedViews: materializedCounts.activeIncremental,
